pyproject.toml
[project]
name = "aimodelchatproductsolution"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "datetime>=6.0",
    "dotenv>=0.9.9",
    "flask>=3.1.2",
    "langchain-core>=1.1.1",
    "langgraph>=1.0.4",
    "requests>=2.32.5",
    "typing>=3.10.0.0",
    "uuid>=1.30",
    "werkzeug>=3.1.4",
    "aiosqlite>=0.21.0",
    "arabic-reshaper>=3.0.0",
    "asyncpg>=0.31.0",
    "chainlit>=2.8.4",
    "docker>=7.1.0",
    "fastmcp>=2.13.1",
    "langchain>=1.0.3",
    "langchain-community>=0.4.1",
    "langchain-mcp-adapters>=0.1.13",
    "langchain-openai>=1.0.2",
    "langchain-tavily>=0.2.13",
    "langgraph-checkpoint-sqlite>=3.0.0",
    "psycopg2>=2.9.11",
    "pydantic>=2.12.3",
    "pyngrok>=7.5.0",
    "sqlalchemy>=2.0.44",
    "toolbox-langchain>=0.5.3",
]

[tool.uv]
package = true

[build-system]
requires = ["setuptools>=64", "wheel"]
#build-backend = "setuptools.build_meta"

[dependency-groups]
dev = [
    "ipykernel>=7.1.0",
    "pandas>=2.3.3",
    "pillow>=12.0.0",
]

[tool.setuptools.packages.find]
where = [".", "agentic_flow"]          
include = ['*']
